Update patch set 1
Patch Set 1:

I'd prefer to fix mount_all too, but I don't think we can (for now):

mount_all <fstab> [ <path> ]* [--<options>]*

So, if we want <fstab> to be optional, it will be ambiguous whether fstab is a <path> or not. This feature seems to have been removed in R, but I guess init needs to remain backwards compatible. 😞

I could go down the road of checking whether the first argument exists as a file or not, but I don't know if that would negatively affect compatibility too.

Is setting a property at "mount_all /explict/path --late" satisfactory for now, and we add the "pathless" mount_all in a future Android version, once the <path>* feature can be deleted?

Patch-set: 1