Update patch set 3
Patch Set 3:

I've thought about this a bit more. Nick, while I agree with your general point that using non-SELinux-aware tools is a Bad Thing™, the whole premise of the restorecon utility is:

1. to change existing file labels
2. in writable directories
3. to conform to the current sepolicy

So, I too see two scenarios of restorecon usage:
1. Newly created files, e.g. in /metadata, /postinstall or /data. Restorecon shouldn't need to be run here, since all file manipulation tools should be SELinux-aware. Yet, there are also
2. Files created with a previous sepolicy revision whose labels are now out of date. This could be a recovery using outdated labels or an update/DSU/whatever process.

In scenario 2, restorecon is too smart by half and respects the top-level sehash, instead of recognizing that in a subdirectory the labels have changed.

Respecting the xattr is merely a performance optimization. As long as this optimization is not 100% dependable, it cannot be trusted. The "force" option is present in libselinux and the only thing this CL changes is to expose it to init.

By the way, OEMs are hacking in this exact feature feature[^1][^2][^3][^4]. The issue is not limited to Sony.

So by failing to ensure that after restorecon is run, the labels in that directory are 100% guaranteed to match the current sepolicy, you create scenarios where vendors are forced to do bad things like manually clearing the sehash xattr.


I see some work being done in [^5]. If that CL resolves the current issue, I will close this topic once and for all.

[^1] Exhibit A: https://github.com/AndroidDumps/sony_j9110_dump/blob/griffin_dsds-user-10-55.1.A.3.149-055001A003014902854466984/vendor/etc/init/hw/init.target.rc#L58
[^2] Exhibit B: https://github.com/C457/android_device_zte_pine/blob/6ce6b505b487868f1c22abf441bce2d8be539c26/rootdir/etc/init.target.rc#L85
[^3] Exhibit C: https://github.com/chemusqui1/android_device_samsung_universal8895-common/blob/18dbfbd8acfb84cfe19af36f69e525bc281b7a12/ramdisk/etc/init.samsung.rc#L304
[^4] Exhibit D-... https://github.com/search?p=12&q=restorecon_recursive+--force&type=Code
[^5] https://android-review.googlesource.com/c/platform/external/selinux/+/1291737

Patch-set: 3
Reviewer: Gerrit User 1421169 <1421169@85c56323-6fa9-3386-8a01-6480fb634889>