Update patch set 3
Patch Set 3:

I have a short comment for this.
Actually, we need to consider FDE case here.
When metadata encryption is on, the content of super block alway looks broken.
In this case, this function needs to return -EINVAL in the same way with the previous.
AFAIK, we should not recover super block in this case, if we do that, instead, we would break the super block. :(

Patch-set: 3
Reviewer: Gerrit User 1651802 <1651802@85c56323-6fa9-3386-8a01-6480fb634889>