Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> I'll defer to smoreland@ too.  Unfortunately he's away for a bit, so I don't think we'll get a response soon.
> 
> +spivack@ too for his thoughts.

Unfortunately, I'm only familiar with the AIDL side of things, which doesn't support multiple versions of interfaces or use fully qualified names. So the only suggestion I can offer is to make sure the function is robust enough to handle the case in which a user inexplicably provides an AIDL service name as input (though I see there is already a TODO for this).

Patch-set: 3