Update patch set 5
Patch Set 5:

After looking through your review, I effectively rewrote the fuzzer to try to attack the problem in a more effective manner. I am now solely focusing on Unwinder, and based a lot of logic off the JIT and Elf tests. This seems to be a good way to get higher coverage (seems to increase even after several million iterations) and generally a more robust approach. I appreciate the information you provided and hope this did a good job of incorporating that feedback. Due to the scale of the changes made I will likely be resolving most, if not all of the existing feedback, as it's effectively a complete rewrite.

Patch-set: 5