Add dmuserd, a daemon to handle dm-user requests
This is less of a skeleton of the daemon that just the smallest code that will
provide a full round-trip in the kernel code.  It's relying on all sorts of
kernel implementation oddities (read/write sizes, for example) to function
right now that will almost certainly change as the kernel code gets fleshed
out, but it does at least cause messages to flow right now.

Test: Running "dmuserd" doesn't panic the kernel.
Signed-off-by: Palmer Dabbelt <palmerdabbelt@google.com>
Change-Id: I4cb69ee8274e6a121ecab0903b3f9a17be30f177