Move build.prop to ./etc subdir
build.prop files are moved to the ./etc subdir. This is to be in-line
with the established practice of having all configuration files
under ./etc subdir. build.prop doesn't need to be an exception.

The old patch (<partition>/build.prop and/or <partition>/default.prop)
is supsported in some unbundled partitions (e.g. vendor, odm, etc.)
because we havet to support those partitions that were built with
previous versions of Android.

Bug: 158735147
Test: device has build.prop under ./etc and is bootable
Change-Id: I26300574c23c78cc4e5bc20f98fc9702838465e1