trusty: rpmb_dev: Reduce likelihood of desync
The mock RPMB can get out of sync with TD storage blocks in the event of
panic or crash. Making this O_SYNC will reduce the likelihood of MAC
mismatch. Since storageproxyd and the mock RPMB will still not commit at
the same time, this will not eliminate the problem, but it will
drastically decrease the window where crashing will bring them out of
sync.

Bug: 160723681
Test: None
Change-Id: Ia16dfc88ceae56e4e9724a42ac0db492c0e707a4