Fix a corner case in ReadUevent() that triggers duplicate firmware loading.
Presently, within ReadUevent(), true is returned for a successful case as
well as for the case where we read an invalid uevent (overflowed buffer)
where the Uevent object is not cleared, and the caller calls the
callback (with the earlier stored uevent object), leading to duplicate
firmware loading.

Uevent uevent;
while (ReadUevent(&uevent)) {
  if (callback(uevent) == ListenerAction::kStop) return;
}

Scenario:
1. Proper Uevent received and callback called and firmware loading triggered
2. Overflowed uevent is received, but still the ReadUevent() returns true
   but the uevent object still has earlier event.
3. callback is again called, leading to duplicate firmware load.

Handle it by adding another return value to let the caller know if the
uevent read is invalid, and caller can choose to ignore it and read further
pending uevents.

Change-Id: I09e80052337fd1495b968dc02ecff5ceb683da18