Initial prototype of COW format, API, and offline converter tool.
This is an initial prototype of the new COW format. It does not have
support for merge sequencing or merge state tracking.

The reader and writer APIs have pure virtual interfaces to support
mocking. The writer and reader both support gz compression. In the
current design it only compresses one block at a time.

Also included is a tool called "vab_convert_payload". This takes in a
payload.bin file (from an OTA package), and converts it into COWs in the
new format. It uses the CowWriter API to do this. This tool can be used
to precisely see the size of a COW relative to an OTA. Since there is
one COW per partition, it also takes an output folder:

    vab_convert_payload payload.bin ./ota-cows/

If the payload is incremental, a source target-files package is needed.
It must be the exact package used to build the OTA, otherwise, the
conversion is likely to fail. Example:

    vab_convert_payload -source_tf target-files-XYZ.zip \
                        payload.bin ./ota-cows/

Bug: 162274240
Test: manual tests
      cow_api_test gtest
Change-Id: Ib9fddc1f210700688e6786917b023dfabd5fb3d9