Remove __ANDROID_API__ guards
__ANDROID_API__ guards are removed in favor of __INTRODUCED_IN macros.

For the NDK clients, the headers are processed by the versioner tool
that wraps APIs with the __ANDROID_API__ guards automatically. So the
hand-written guards are not necessary.

For the non-NDK clients, the guards have been preventing us from making
a build system change to let __ANDROID_API__ track the min_sdk_version
property instead of the sdk_version property. With the build system
change, __ANDROID_API__ will suddenly drop for the native modules where
min_sdk_version < sdk_version, which is the case when the modules are
included in APEXes. As a result, some new APIs will be unavailable at
build-time. Dropping the hand-written guards fixes the problem.

Bug: 163288375
Test: m
Change-Id: If1cc6b9af410f536abe6d777c22711209fa76530