Change PropertyMap::load to return owned object
Currently, the caller is responsible for managing the lifecycle of the
returned outMap when calling PropertyMap::load. However, the fact that
the function call allocates new memory is not obvious from the function
signature.
Refactor the function to return Result<unique_ptr<>> to make it less
errorprone.

The input code is the only customer of PropertyMap.

Bug: 163171599
Test: atest inputflinger_tests
Change-Id: I316084886c3f09a1776fdb449d2f03d0563b66c1