snapuserd: Be explicit about making the arrays
I don't actually understand this one: the old code appears to allocate an arroy
for the first call to make_unique<char>(size_t), but then a single element for
the re-allocation calls in the loop which triggers a buffer overflow.  My
attempts to duplicate this in a simple example aren't working out, but I figure
it's better to just be explicit about what we want here.

Test: manual
Signed-off-by: Palmer Dabbelt <palmerdabbelt@google.com>
Change-Id: Iade9ea108b367aafacf28a6d308ee98ad5f000cf