Make thread priority API C++ only
In https://android-review.googlesource.com/c/platform/system/core/+/1224544
we introduced a default argument to androidSetThreadPriority().
This made the function no longer work as a C API, and the header
file as a whole no longer usable from C source.

We now move this function to a C++ only section of the header, so
the header itself can still be included by C code.  For symmetry,
we also move androidGetThreadPriority().

Test: TreeHugger
Bug: 165009705
Change-Id: Ib595e27d86144fd66abf176198a90187de950114