Add benchmark for property mapping
This benchmarks mapping property prefixes to property contexts with
two algorithms: the 'Legacy' method used before Android P and the
'Trie' used afterwards (the code in this directory).

It uses input mappings from both Oreo and the latest in AOSP ('S').
Note that there is nearly a 10x increase in the number of mappings in
S as there was in Oreo, which was predicted when the trie was
designed.

Results on cuttlefish:
-----------------------------------------------------------
Benchmark                 Time             CPU   Iterations
-----------------------------------------------------------
LegacyLookupOreo     515159 ns       513838 ns         1346
LegacyLookupS       3917729 ns      3907313 ns          178
TrieLookupOreo       217533 ns       216963 ns         3216
TrieLookupS          429449 ns       428300 ns         1623

The results show that the legacy look up uses 7.6x more CPU time to
handle the number of mappings added through S, whereas the Trie lookup
uses less than 2x more CPU time, showing that the trie scales better
with added mappings.

Test: run this benchmark
Change-Id: Ia3727090f3fa0cd31e7a8277194ffb74482c795d