libsnapshot: Partially implement OpenSnapshotWriter.
Implement OpenSnapshotWriter for non-compressed Virtual A/B. This is
done by adding an OnlineKernelCowWriter class, which forwards all writes
to a dm-snapshot block device. In this mode, we do not support the
AddCopy() operation, since we don't track the original block device.

To test the new call, vts_libsnapshot_test has been modified to use
OpenSnapshotWriter instead of MapUpdateSnapshot. Internally, the latter
calls the former, so the underlying code is still being tested.

As part of this change, all consumers of libsnapshot must now link to
libsnapshot_cow.

Bug: 168554689
Test: vts_libsnapshot_test
Change-Id: Ieedfadc557833c1e0540922aabc6e95c80266a64