libunwindstack: Use FDE to unwind x86/x86_64 sigreturn
For x86 and x86_64, use the DWARF FDE in libc.so for __restore and
__restore_rt, rather than detecting the trampoline instructions.

For a signal frame, DwarfSectionImpl<AddressType>::Eval needs to
continue the unwinding even if PC is zero. A zero PC means that the
program has crashed, and we should try to recover the real PC using the
return address on the stack or LR. This behavior is tested by
remote_through_signal_with_invalid_func and
remote_through_signal_sa_siginfo_with_invalid_func.

Test: libunwindstack_test
Change-Id: I4655b070028fd984345311a5e743796f8c30ed36