libsnapshot: Group CowOperations into clusters
This introduces clusters of CowOperations to the Cow Format, improving
reader performance by allowing many operations to be read at once.

When the header contains a cluster size, operations will be grouped into
clusters of that size, with the final op being a special CowClusterOp
which has the offset to the next cluster. Each cluster is immediately
followed by its data, with the footer found at the end of the file.

The cluster size must be a multiple of sizeof(CowOperation). If nonzero,
there must be room for at least 2 CowOperations to give room for the
CowClusterOp at the end.

Change-Id: I2855288958d56569d4078f377ef0b8c01ca5c6e7
Bug: 172026020
Test: cow_api_test