init: Add an selinux transition for snapuserd.
With compressed VAB updates, it is not possible to mount /system without
first running snapuserd, which is the userspace component to the dm-user
kernel module. This poses a problem because as soon as selinux
enforcement is enabled, snapuserd (running in a kernel context) does not
have access to read and decompress the underlying system partition.

To account for this, we split SelinuxInitialize into three steps:

First, sepolicy is loaded as it was before.

Second, we re-launch snapuserd and re-create affected device-mapper
tables, to re-route requests to the new daemon. Since the daemon is
running in the snapuserd context, rather than the kernel context, it has
the correct permissions. Note that to achieve this, we first have to
restorecon /dev/block, /dev/dm-user, /dev/socket, and snapuserd itself.
Also, this requires adding an selinux transition from kernel to snapuserd.

Finally, we set enforcing mode. At this point the original
kernel-privileged snapuserd has been killed.

The newly introduced second step generates audits, but not denials. The
restorecons are normally expected from the init context and not kernel,
but they are run before enforcement is enabled.

Bug: 173476209
Test: full OTA with VABC applies and boots
Change-Id: Ie4e0f5166b01c31a6f337afc26fc58b96217604e