trusty: Add library for fuzzing replay tools
Adds a static library, libtrusty_coverage_tool, which provides a
standalone main function for libfuzzer fuzzers. This main function takes
a set of input files (e.g., the corpus after fuzzing) on the command
line and sends each to the LLVMFuzzerTestOneInput entry point in turn.
The tool communicates with the Trusty coverage server and saves the full
set of covered PCs to a sancov file after all input has been sent.

Bug: 175221942
Test: make libtrusty_coverage_tool
Change-Id: Id7a0ca43ec8b5803d501848e2673e5ad013e7808