adb remount -R checks DSU running and enabled state
If executing `adb remount -R` and DSU is running but disabled, then
don't auto reboot upon remount completion.
This is because rebooting in this state (DSU running and disabled) would
bring the device out of the DSU guest system and back to the host
system, which probably isn't what the user is expecting.
What the user would want to do is first
$ adb shell gsi_tool enable -s
to ensure the device would still be in the DSU guest system after
reboot, and then do
$ adb remount -R
to setup the remount environment and reboot.

Also sort the header include order.

Bug: 165925766
Test: Within a DSU guest system, DSU disabled,
  adb remount -R => doesn't reboot and emit a warning
Test: Within a DSU guest system, DSU enabled,
  adb remount -R => reboot and no warning
Test: adb-remount-test.sh within DSU guest system
Change-Id: I72a7a568e985b183d357ae6e1a7d0113e9921200