Create apex reserved dir for update_engine
When installing an OTA, update_engine needs to reserve some space on
filesystem by writing to a specific directory(/data/apex/reserved/),
therefore we need to create this dir on system start up.

We are also pro-actively enabling encryption on this directory so that
we can create hard links to /data/apex/decompressed. This will be needed
when we start decompresssing capex from post-install script before
reboot and on boot we can then simply hard link to these files.

Test: th
Bug: 172911822

Change-Id: Ia6a63efcedcfdad9817ba88b54f96683d34df6ce