ueventd: determines the order of firmware path based on the number of times requested
The current ueventd is scanning /apex as well, so it can load updatable firmware.
However, if the loaded firmware from /apex is invalid, the kernel continues to retry,
but the rollback conditions do not work.
If the firmware is installed on both /system (or /vendor) and /apex,
fallback can be considered,
so it scans /apex first in normal cases
and the request count will decide the firmware path to activate the fallback in case of a problem.

Test: build pass
Test: boot on device after installing APEX with firmware
Test: install invalid firmware and check fallback
case #1) install same firmware both /vendor and /apex -> load firmware from /apex
case #2) install new firmware to /apex -> load firmware from /apex
case #3) install new invalid firmware to /apex -> load firmware from /apex first
         and reload default firmware from /vendor

Signed-off-by: ByungSoo Kim <bs1770.kim@samsung.com>
Change-Id: Ic55efafecaa6baebf695f9da0866a9d1e28d6945