Grab input event before triggering TUI session
If the power key is pressed down before TUI session is started while the
input events are not grabbed, it could trigger the non-secure world to
trigger the system to enter suspend which causes the display state being
changed. The changed display state might lead to display state not sync
as the secure world will also change the display state during the TUI
session.

Bug: 148421469
Test: TUI VTS/CTS pass
Change-Id: I255934a3acae65d8a4b4a3e44120257e362ba9df