Fix userspace-reboot failure when backing_dev exists but zram not swapped on
'/sys/block/zram0/backing_dev' will exist even if zram is not swapped on in some devices. And there is no reason to ensure that zram is swapped on if '/sys/block/zram0/backing_dev' exists. So, if we want to kill backing_dev during userspace reboot, wo check if zram is swapped on first.

Signed-off-by: luwei9 <hellmosv60@gmail.com>
Change-Id: Ie707c2e61b1f757d9e67e2a6a11249f7d43c2027