[MTE] Add a HWASan-style tag dump to tombstones.
We already dump the tags in the regigster dump section by appending the
tag to the memory address. You only get 2 granules before each register
and 13 after.

The HWASan-style tag dump is extremely useful for debugging, as it gives
a pretty comprehensive overview of the memory subsystem. It also
provides enough context bytes (256) to give you a reasonable intuition
about a particular bug.

The tag dump shows up only if PTRACE_PEEKTAGS returns at least one value
in the 256 requested. If the start of end of the region is untagged,
it's omitted. The tag dump looks like this:

Change-Id: I01934f4cbd3464feae56eddf1a373f55af681333

---
Memory tags around the fault address (0x7000072f6d153d0), one tag per 16 bytes:
      0x72f6d14b00: 7  7  7  7  0  0  2  2  2  2  2  2  2  0  0  3
      0x72f6d14c00: 3  3  3  3  3  3  0  0  4  4  4  4  4  4  4  0
      0x72f6d14d00: 0  9  9  9  9  9  9  9  0  0  8  8  8  8  8  8
      0x72f6d14e00: 8  0  0  5  5  5  5  5  5  5  0  0  a  a  a  a
      0x72f6d14f00: a  a  a  0  0  7  7  7  7  7  7  7  0  0  e  e
      0x72f6d15000: e  e  e  e  e  0  0  b  b  b  b  b  b  b  0  0
      0x72f6d15100: 0  0  0  0  0  0  0  0  0  f  f  f  f  f  f  f
      0x72f6d15200: 0  0  8  8  8  8  8  8  8  0  0  f  f  f  f  f
    =>0x72f6d15300: f  f  0  0  e  e  e  e  e  e  e  0  0 [b] b  b
      0x72f6d15400: b  b  b  b  0  0  c  c  c  c  c  c  c  0  0  f
      0x72f6d15500: f  f  f  f  f  f  0  0  8  8  8  8  8  8  8  0
      0x72f6d15600: 0  d  d  d  d  d  d  d  0  0  2  2  2  2  2  2
      0x72f6d15700: 2  0  0  b  b  b  b  b  b  b  0  0  e  e  e  e
      0x72f6d15800: e  e  e  0  0  f  f  f  f  f  f  f  0  0  e  e
      0x72f6d15900: e  e  e  e  e  0  0  0  0  0  0  0  0  0  0  0
      0x72f6d15a00: a  a  a  a  a  a  a  0  0  0  0  0  0  0  0  0
---

Bug: 183992164
Test: atest debuggerd_test on MTE+QEMU and sunfish.

Change-Id: I8d5842e4803ca30b407e866c99eef56f2cb36600